+++ pages/favorites/components/Players.tsx
